move PolymorphicSlots into Slotable
