+++ lib/view_component/base.rb
@@ -6,7 +6,6 @@ require "view_component/collection"
 require "view_component/compile_cache"
 require "view_component/compiler"
 require "view_component/config"
-require "view_component/polymorphic_slots"
 require "view_component/preview"
 require "view_component/slotable"
 require "view_component/translatable"
@@ -29,7 +28,6 @@ module ViewComponent
       attr_writer :config
     end
 
-    include ViewComponent::PolymorphicSlots
     include ViewComponent::Slotable
     include ViewComponent::Translatable
     include ViewComponent::WithContentHelper
